+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@
 ## v2.1.3 (2022-02-02)
 
 - Documentation improvements and small changes;
+- Fixed bug in test file `tests/utils.go` related to not properly counting months and years when testing;
 
 ----